backend/requirements.txt
# TenderWatch SaaS – Backend Dependencies

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pyjwt>=2.8.0

# Supabase
supabase>=2.4.0

# Scheduler
apscheduler>=3.10.0
httpx>=0.27.0

# AI
anthropic>=0.40.0

# Scraping (from TenderWatch v2)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0

# PDF
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Utils
python-dotenv>=1.0.0
urllib3>=2.0.0
